backend/requirements.txt
# requirements.txt (Python backend)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
mysql-connector-python==8.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2